Deploy marriage abroad
